chore(catalog): add ClusterOperator status for catalog operator
previously we were only reporting OLM status
